CE-373: migrate team manual
What?
renaming the concourse pipeline, jobs and tasks as the this team-manual
lambda function was done as part of firebreak but it now a live thing.

Why?
We moved the team-manual from the PaaS to the lambda function because we
have already google auth working with it on the lambda function.

Tested?
Yes, it has been. This new pipeline is already live on our concourse ci
instance and it runs cleanly. https://cd.gds-reliability.engineering/teams/cybersecurity-tools/pipelines/cyber-team-manual-lambda-deploy

@ronocg <conor.glynn@digital.cabinet-office.gov.uk>